Created a nice widget and the YAML code to open up "Path" through a filechooser (and read it up as string). Also, corrected the dashboard yamls since they weren't given the new name for the Abstract Scenario (the scenario formerly known as Prototype)

+++ src/main/java/uk/ac/ox/oxfish/gui/FishGUI.java
@@ -198,7 +198,7 @@ public class FishGUI extends GUIState{
         //change width and height to keep correct geographical ratio
         double width;
         double height;
-        double heightToWidthRatio = state.getRasterBathymetry().getGridHeight()/state.getRasterBathymetry().getGridWidth();
+        double heightToWidthRatio = ((double) state.getRasterBathymetry().getGridHeight())/state.getRasterBathymetry().getGridWidth();
         if(heightToWidthRatio >= 1)
         {
             width = MIN_DIMENSION;
